Reorganize ScriptCompiler code
