+++ Server/Server.csproj
@@ -138,7 +138,6 @@
     <Compile Include=".\QuestArrow.cs" />
     <Compile Include=".\Race.cs" />
     <Compile Include=".\Region.cs" />
-    <Compile Include=".\ScriptCompiler.cs" />
     <Compile Include=".\Sector.cs" />
     <Compile Include=".\SecureTrade.cs" />
     <Compile Include=".\Serial.cs" />
@@ -158,9 +157,6 @@
     <Compile Include=".\Utility.cs" />
     <Compile Include=".\VirtueInfo.cs" />
     <Compile Include=".\World.cs" />
-    <Compile Include="CachedCompilerBackend.cs" />
-    <Compile Include="CodeDomCompilerBackend.cs" />
-    <Compile Include="Compiler.cs" />
     <Compile Include="Config.cs" />
     <Compile Include="Gumps\Gump.cs" />
     <Compile Include="Gumps\GumpAlphaRegion.cs" />
@@ -188,12 +184,18 @@
     <Compile Include="Gumps\KRGumpImage.cs" />
     <Compile Include="Gumps\KRGumpLabel.cs" />
     <Compile Include="Gumps\RelayInfo.cs" />
-    <Compile Include="ICompilerBackend.cs" />
     <Compile Include="Persistence\Persistence.cs" />
     <Compile Include="Berserk.cs" />
     <Compile Include="PromptExtensions.cs" />
     <Compile Include="Random.cs" />
     <Compile Include="Items\VirtualCheck.cs" />
+    <Compile Include="Scripts\CachedCompilerBackend.cs" />
+    <Compile Include="Scripts\CodeDomCompilerBackend.cs" />
+    <Compile Include="Scripts\Compiler.cs" />
+    <Compile Include="Scripts\ICompilerBackend.cs" />
+    <Compile Include="Scripts\ScriptCompiler.cs" />
+    <Compile Include="Scripts\TypeCache.cs" />
+    <Compile Include="Scripts\TypeTable.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="servuo.ico" />
